Solves issues from Trac #3682.
In the scenario info window:
  - Coordinates are shown in RD, not WGS84
  - Intervals are shown as '2 d 12:00', not as 2.5
  - Values like '-999' are treated as null values

Created a "coordinates.py" that contains functions relating to coordinates.

Made two methods in views_infowindow.py, find_imported_value tries to
find a value in the database, given the inputfield it would have come
from in case of an import, and display_string returns a string version
of the value, depending on the type of that same inputfield.

Wrote some tests for both.